Split tests
- Split tests to make it easier to dev features for item types with always having to run all tests
