+++ README.md
@@ -287,26 +287,29 @@ You can see our progress in the table below. We are hoping to add tests in each 
 not too concerned about missing anything as we still have all our tests in Postman, we won't disable our test monitor until 
 our local test suite is as complete as the Postman request test suite.
 
-| Controller          | Action   | View     | 
-|:--------------------|:---------|:---------|
-| Authentication      | 35 Tests | 3 Tests  |
-| Category            | 21 Tests | 27 Tests |
-| Currency            | Non yet* | Non yet* |
-| ItemCategory        | Non yet* | Non yet* |
-| Item                | 45 Tests | 39 Tests |
-| ItemData            | Non yet* | Non yet* |
-| ItemLog             | Non yet* | Non yet* |
-| ItemPartialTransfer | Non yet* | Non yet* |
-| ItemSubcategory     | Non yet* | Non yet* |
-| ItemTransfer        | Non yet* | Non yet* |
-| ItemType            | Non yet* | 7 Tests  |
-| PermittedUser       | 4 Tests  | 2 Tests  |
-| Queue               | Non yet* | Non yet* |
-| Request             | Non yet* | Non yet* |
-| Resource            | 24 Tests | 27 Tests |
-| ResourceType        | 23 Tests | 26 Tests |
-| Subcategory         | 21 Tests | 23 Tests |
-| **Total tests**     | **173**  | **154**  |
+| Controller               | Action   | View     | 
+|:-------------------------|:---------|:---------|
+| Authentication           | 35 Tests | 3 Tests  |
+| Category                 | 21 Tests | 27 Tests |
+| Currency                 | Non yet* | Non yet* |
+| ItemCategory             | Non yet* | Non yet* |
+| Item (Allocated Expense) | 9 Tests  | 9 Tests  |
+| Item (Budget)            | 12 Tests | 9 Tests  |
+| Item (Budget Pro)        | 12 Tests | 9 Tests  |
+| Item (Game)              | 12 Tests | 12 Tests |
+| ItemData                 | Non yet* | Non yet* |
+| ItemLog                  | Non yet* | Non yet* |
+| ItemPartialTransfer      | Non yet* | Non yet* |
+| ItemSubcategory          | Non yet* | Non yet* |
+| ItemTransfer             | Non yet* | Non yet* |
+| ItemType                 | Non yet* | 7 Tests  |
+| PermittedUser            | 4 Tests  | 2 Tests  |
+| Queue                    | Non yet* | Non yet* |
+| Request                  | Non yet* | Non yet* |
+| Resource                 | 24 Tests | 27 Tests |
+| ResourceType             | 23 Tests | 26 Tests |
+| Subcategory              | 21 Tests | 23 Tests |
+| **Total tests**          | **173**  | **154**  |
 
 *Non yet does not mean there are no tests, it just means there are no PHPUnit tests. There are over 2000 tests in 
 a private Postman collection, I'm slowing transferring them locally and expanding the test suite.
