+++ tests/Action/Http/Controllers/ItemAllocatedExpenseTest.php
@@ -1,0 +1,194 @@
+<?php
+
+namespace Http\Controllers;
+
+use App\User;
+use Illuminate\Support\Str;
+use Tests\TestCase;
+
+final class ItemAllocatedExpenseTest extends TestCase
+{
+    /** @test */
+    public function createAllocatedExpenseItemFailsCurrencyIdInvalid(): void
+    {
+        $this->actingAs(User::find(1));
+
+        $resource_type_id = $this->createAllocatedExpenseResourceType();
+        $resource_id = $this->createAllocatedExpenseResource($resource_type_id);
+
+        $response = $this->createItem(
+            $resource_type_id,
+            $resource_id,
+            [
+                'name' => $this->faker->text(200),
+                'description' => $this->faker->text(200),
+                'effective_date' => $this->faker->date(),
+                'currency_id' => 'epMqeYqPkp',
+                'total' => $this->randomMoneyValue(),
+            ]
+        );
+
+        $response->assertStatus(422);
+    }
+
+    /** @test */
+    public function createAllocatedExpenseItemFailsNoNameInPayload(): void
+    {
+        $this->actingAs(User::find(1));
+
+        $resource_type_id = $this->createAllocatedExpenseResourceType();
+        $resource_id = $this->createAllocatedExpenseResource($resource_type_id);
+
+        $response = $this->createItem(
+            $resource_type_id,
+            $resource_id,
+            [
+                'description' => $this->faker->text(200),
+                'effective_date' => $this->faker->date(),
+                'currency_id' => $this->currency['GBP'],
+                'total' => $this->randomMoneyValue(),
+            ]
+        );
+
+        $response->assertStatus(422);
+    }
+
+    /** @test */
+    public function createAllocatedExpenseItemFailsNoPayload(): void
+    {
+        $this->actingAs(User::find(1));
+
+        $resource_type_id = $this->createAllocatedExpenseResourceType();
+        $resource_id = $this->createAllocatedExpenseResource($resource_type_id);
+
+        $response = $this->createItem(
+            $resource_type_id,
+            $resource_id,
+            [
+            ]
+        );
+
+        $response->assertStatus(422);
+    }
+
+    /** @test */
+    public function createAllocatedExpenseItemSuccess(): void
+    {
+        $this->actingAs(User::find(1));
+
+        $resource_type_id = $this->createAllocatedExpenseResourceType();
+        $resource_id = $this->createAllocatedExpenseResource($resource_type_id);
+
+        $response = $this->createItem(
+            $resource_type_id,
+            $resource_id,
+            [
+                'name' => $this->faker->text(200),
+                'description' => $this->faker->text(200),
+                'effective_date' => $this->faker->date(),
+                'currency_id' => $this->currency['GBP'],
+                'total' => $this->randomMoneyValue(),
+            ]
+        );
+
+        $response->assertStatus(201);
+        $this->assertJsonMatchesAllocatedExpenseItemSchema($response->content());
+    }
+
+    /** @test */
+    public function deleteAllocatedExpenseItemFailsIdNotFound(): void
+    {
+        $this->actingAs(User::find(1));
+
+        $resource_type_id = $this->createAllocatedExpenseResourceType();
+        $resource_id = $this->createAllocatedExpenseResource($resource_type_id);
+        $item_id = '1234asdffgd';
+
+        $response = $this->deleteItem(
+            $resource_type_id,
+            $resource_id,
+            $item_id,
+        );
+
+        $response->assertStatus(403);
+    }
+
+    /** @test */
+    public function deleteAllocatedExpenseItemSuccess(): void
+    {
+        $this->actingAs(User::find(1));
+
+        $resource_type_id = $this->createAllocatedExpenseResourceType();
+        $resource_id = $this->createAllocatedExpenseResource($resource_type_id);
+        $item_id = $this->createAllocatedExpenseItem($resource_type_id,$resource_id);
+
+        $response = $this->deleteItem(
+            $resource_type_id,
+            $resource_id,
+            $item_id,
+        );
+
+        $response->assertStatus(204);
+    }
+
+    /** @test */
+    public function updateAllocatedExpenseItemFailsNonExistentField(): void
+    {
+        $this->actingAs(User::find(1));
+
+        $resource_type_id = $this->createAllocatedExpenseResourceType();
+        $resource_id = $this->createAllocatedExpenseResource($resource_type_id);
+        $item_id = $this->createAllocatedExpenseItem($resource_type_id, $resource_id);
+
+        $response = $this->updateItem(
+            $resource_type_id,
+            $resource_id,
+            $item_id,
+            [
+                'does_not_exist' => $this->faker->text(25)
+            ]
+        );
+
+        $response->assertStatus(400);
+    }
+
+    /** @test */
+    public function updateAllocatedExpenseItemFailsNoPayload(): void
+    {
+        $this->actingAs(User::find(1));
+
+        $resource_type_id = $this->createAllocatedExpenseResourceType();
+        $resource_id = $this->createAllocatedExpenseResource($resource_type_id);
+        $item_id = $this->createAllocatedExpenseItem($resource_type_id, $resource_id);
+
+        $response = $this->updateItem(
+            $resource_type_id,
+            $resource_id,
+            $item_id,
+            []
+        );
+
+        $response->assertStatus(400);
+    }
+
+    /** @test */
+    public function updateAllocatedExpenseItemSuccess(): void
+    {
+        $this->actingAs(User::find(1));
+
+        $resource_type_id = $this->createAllocatedExpenseResourceType();
+        $resource_id = $this->createAllocatedExpenseResource($resource_type_id);
+        $item_id = $this->createAllocatedExpenseItem($resource_type_id, $resource_id);
+
+        $response = $this->updateItem(
+            $resource_type_id,
+            $resource_id,
+            $item_id,
+            [
+                'name' => $this->faker->text(25)
+            ]
+        );
+
+        $response->assertStatus(204);
+    }
+}
