+++ tests/Action/Http/Controllers/ItemBudgetTest.php
@@ -1,0 +1,280 @@
+<?php
+
+namespace Http\Controllers;
+
+use App\User;
+use Illuminate\Support\Str;
+use Tests\TestCase;
+
+final class ItemBudgetTest extends TestCase
+{
+    /** @test */
+    public function createBudgetItemFailsAmountNotFormattedCorrectly(): void
+    {
+        $this->actingAs(User::find(1));
+
+        $resource_type_id = $this->createBudgetResourceType();
+        $resource_id = $this->createBudgetResource($resource_type_id);
+
+        $response = $this->createItem(
+            $resource_type_id,
+            $resource_id,
+            [
+                'name' => $this->faker->text(200),
+                'account' => Str::uuid()->toString(),
+                'description' => $this->faker->text(200),
+                'amount' => number_format($this->faker->randomFloat(2, 0.01, 99999999999.99), 2, '.', ','),
+                'currency_id' => $this->currency['GBP'],
+                'category' => 'income',
+                'start_date' => $this->faker->date(),
+                'frequency' => json_encode(['type'=>'monthly', 'day'=>null, 'exclusions' => []], JSON_THROW_ON_ERROR),
+            ]
+        );
+
+        $response->assertStatus(422);
+    }
+
+    /** @test */
+    public function createBudgetItemFailsCategoryInvalid(): void
+    {
+        $this->actingAs(User::find(1));
+
+        $resource_type_id = $this->createBudgetResourceType();
+        $resource_id = $this->createBudgetResource($resource_type_id);
+
+        $response = $this->createItem(
+            $resource_type_id,
+            $resource_id,
+            [
+                'name' => $this->faker->text(200),
+                'account' => Str::uuid()->toString(),
+                'description' => $this->faker->text(200),
+                'amount' => $this->randomMoneyValue(),
+                'currency_id' => $this->currency['GBP'],
+                'category' => 'not-income',
+                'start_date' => $this->faker->date(),
+                'frequency' => json_encode(['type'=>'monthly', 'day'=>null, 'exclusions' => []], JSON_THROW_ON_ERROR),
+            ]
+        );
+
+        $response->assertStatus(422);
+    }
+
+    /** @test */
+    public function createBudgetItemFailsCurrencyInvalid(): void
+    {
+        $this->actingAs(User::find(1));
+
+        $resource_type_id = $this->createBudgetResourceType();
+        $resource_id = $this->createBudgetResource($resource_type_id);
+
+        $response = $this->createItem(
+            $resource_type_id,
+            $resource_id,
+            [
+                'name' => $this->faker->text(200),
+                'account' => Str::uuid()->toString(),
+                'description' => $this->faker->text(200),
+                'amount' => $this->randomMoneyValue(),
+                'currency_id' => 'epMqeYqPko',
+                'category' => 'income',
+                'start_date' => $this->faker->date(),
+                'frequency' => json_encode(['type'=>'monthly', 'day'=>null, 'exclusions' => []], JSON_THROW_ON_ERROR),
+            ]
+        );
+
+        $response->assertStatus(422);
+    }
+
+    /** @test */
+    public function createBudgetItemFailsFrequencyJsonInvalid(): void
+    {
+        $this->actingAs(User::find(1));
+
+        $resource_type_id = $this->createBudgetResourceType();
+        $resource_id = $this->createBudgetResource($resource_type_id);
+
+        $response = $this->createItem(
+            $resource_type_id,
+            $resource_id,
+            [
+                'name' => $this->faker->text(200),
+                'account' => Str::uuid()->toString(),
+                'description' => $this->faker->text(200),
+                'amount' => $this->randomMoneyValue(),
+                'currency_id' => $this->currency['GBP'],
+                'category' => 'income',
+                'start_date' => $this->faker->date(),
+                'frequency' => '{"field"=>true}',
+            ]
+        );
+
+        $response->assertStatus(422);
+    }
+
+    /** @test */
+    public function createBudgetItemFailsNoNameInPayload(): void
+    {
+        $this->actingAs(User::find(1));
+
+        $resource_type_id = $this->createBudgetResourceType();
+        $resource_id = $this->createBudgetResource($resource_type_id);
+
+        $response = $this->createItem(
+            $resource_type_id,
+            $resource_id,
+            [
+                'account' => Str::uuid()->toString(),
+                'description' => $this->faker->text(200),
+                'amount' => $this->randomMoneyValue(),
+                'currency_id' => $this->currency['GBP'],
+                'category' => 'income',
+                'start_date' => $this->faker->date(),
+                'frequency' => json_encode(['type'=>'monthly', 'day'=>null, 'exclusions' => []], JSON_THROW_ON_ERROR),
+            ]
+        );
+
+        $response->assertStatus(422);
+    }
+
+    /** @test */
+    public function createBudgetItemFailsNoPayload(): void
+    {
+        $this->actingAs(User::find(1));
+
+        $resource_type_id = $this->createBudgetResourceType();
+        $resource_id = $this->createBudgetResource($resource_type_id);
+
+        $response = $this->createItem(
+            $resource_type_id,
+            $resource_id,
+            []
+        );
+
+        $response->assertStatus(422);
+    }
+
+    /** @test */
+    public function createBudgetItemSuccess(): void
+    {
+        $this->actingAs(User::find(1));
+
+        $resource_type_id = $this->createBudgetResourceType();
+        $resource_id = $this->createBudgetResource($resource_type_id);
+
+        $response = $this->createItem(
+            $resource_type_id,
+            $resource_id,
+            [
+                'name' => $this->faker->text(200),
+                'account' => Str::uuid()->toString(),
+                'description' => $this->faker->text(200),
+                'amount' => $this->randomMoneyValue(),
+                'currency_id' => $this->currency['GBP'],
+                'category' => 'income',
+                'start_date' => $this->faker->date(),
+                'frequency' => json_encode(['type'=>'monthly', 'day'=>null, 'exclusions' => []], JSON_THROW_ON_ERROR),
+            ]
+        );
+
+        $response->assertStatus(201);
+        $this->assertJsonMatchesBudgetItemSchema($response->content());
+    }
+
+    /** @test */
+    public function deleteBudgetItemFailsIdNotFound(): void
+    {
+        $this->actingAs(User::find(1));
+
+        $resource_type_id = $this->createBudgetResourceType();
+        $resource_id = $this->createBudgetResource($resource_type_id);
+        $item_id = '1234asdffgd';
+
+        $response = $this->deleteItem(
+            $resource_type_id,
+            $resource_id,
+            $item_id,
+        );
+
+        $response->assertStatus(403);
+    }
+
+    /** @test */
+    public function deleteBudgetItemSuccess(): void
+    {
+        $this->actingAs(User::find(1));
+
+        $resource_type_id = $this->createBudgetResourceType();
+        $resource_id = $this->createBudgetResource($resource_type_id);
+        $item_id = $this->createBudgetItem($resource_type_id, $resource_id);
+
+        $response = $this->deleteItem(
+            $resource_type_id,
+            $resource_id,
+            $item_id,
+        );
+
+        $response->assertStatus(204);
+    }
+
+    /** @test */
+    public function updateBudgetItemFailsNonExistentField(): void
+    {
+        $this->actingAs(User::find(1));
+
+        $resource_type_id = $this->createBudgetResourceType();
+        $resource_id = $this->createBudgetResource($resource_type_id);
+        $item_id = $this->createBudgetItem($resource_type_id, $resource_id);
+
+        $response = $this->updateItem(
+            $resource_type_id,
+            $resource_id,
+            $item_id,
+            [
+                'does_not_exist' => $this->faker->text(25)
+            ]
+        );
+
+        $response->assertStatus(400);
+    }
+
+    /** @test */
+    public function updateBudgetItemFailsNoPayload(): void
+    {
+        $this->actingAs(User::find(1));
+
+        $resource_type_id = $this->createBudgetResourceType();
+        $resource_id = $this->createBudgetResource($resource_type_id);
+        $item_id = $this->createBudgetItem($resource_type_id, $resource_id);
+
+        $response = $this->updateItem(
+            $resource_type_id,
+            $resource_id,
+            $item_id,
+            []
+        );
+
+        $response->assertStatus(400);
+    }
+
+    /** @test */
+    public function updateBudgetItemSuccess(): void
+    {
+        $this->actingAs(User::find(1));
+
+        $resource_type_id = $this->createBudgetResourceType();
+        $resource_id = $this->createBudgetResource($resource_type_id);
+        $item_id = $this->createBudgetItem($resource_type_id, $resource_id);
+
+        $response = $this->updateItem(
+            $resource_type_id,
+            $resource_id,
+            $item_id,
+            [
+                'name' => $this->faker->text(25)
+            ]
+        );
+
+        $response->assertStatus(204);
+    }
+}
